+++ source/util/aarot.c
@@ -112,13 +112,13 @@ _aa_rotate_bitmap (BITMAP *_src, BITMAP *_dst, int _x, int _y, fixed _angle,
 
 	sw <<= aa_BITS;
 	dsx = sw / dw;
-	if (dsx < aa_SIZE)
+	if ((unsigned int)dsx < aa_SIZE)
 		dsx = aa_SIZE;
 	sw -= dsx;
 
 	sh <<= aa_BITS;
 	dsy = sh / dh;
-	if (dsy < aa_SIZE)
+	if ((unsigned int)dsy < aa_SIZE)
 		dsy = aa_SIZE;
 	sh -= dsy;
 
@@ -127,9 +127,9 @@ _aa_rotate_bitmap (BITMAP *_src, BITMAP *_dst, int _x, int _y, fixed _angle,
 	/* Avoid overflow.  */
 	if (num > aa_MAX_NUM)
 	{
-		if (dsx > aa_MAX_SIZE)
+		if ((unsigned int)dsx > aa_MAX_SIZE)
 			dsx = aa_MAX_SIZE;
-		if (dsy > aa_MAX_SIZE)
+		if ((unsigned int)dsy > aa_MAX_SIZE)
 			dsy = aa_MAX_SIZE;
 		num = dsx * dsy;
 	}
